core/pyproject.toml
[project]
name = "Cheshire-Cat"
description = "Production ready AI assistant framework"
version = "1.9.2"
requires-python = ">=3.10"
license = { file = "LICENSE" }
authors = [
    { name = "Piero Savastano" },
    { name = "Cheshire Cat AI Contributors" },
]
dependencies = [
    "gunicorn==20.1.0",
    "uvicorn[standard]==0.20.0",
    "python-multipart==0.0.6",
    "pydantic==2.4.2",
    "fastapi==0.115.8",
    "scalar-fastapi==1.0.3",
    "websockets==15.0",
    "PyJWT==2.8.0",
    "bcrypt==4.1.3",
    "Jinja2==3.1.4",
    "pandas==1.5.3",
    "scikit-learn==1.2.1",
    "qdrant_client==1.13.3",
    "langchain==0.2.1",
    "langchain-community==0.2.1",
    "langchain-ollama==0.1.3",
    "langchain-openai==0.1.7",
    "langchain-anthropic==0.1.23",
    "langchain-google-genai==1.0.8",
    "langchain-google-genai[images]",
    "langchain-cohere==0.1.5",
    "huggingface-hub>=0.25.1",
    "beautifulsoup4==4.12.3",
    "pdfminer.six==20221105",
    "unstructured==0.12.6",
    "tiktoken==0.7.0",
    "tomli",
    "loguru==0.7.2",
    "text_generation==0.6.1",
    "tinydb==4.8.0",
    "python-slugify==8.0.1",
    "autopep8",
    "pylint",
    "perflint",
    "pylint-actions",
    "pytest",
    "pytest-asyncio",
    "httpx==0.27.2",
    "fastembed==0.3.6",
    "rapidfuzz==3.6.1",
    "APScheduler==3.10.4",
    "ruff==0.4.7",
    "aiofiles==24.1.0",
    "psycopg2-binary==2.9.11"
]

[tool.coverage.run]
source = ["cat"]

[tool.pylint]
max-line-length = 180
disable = [
    "C0301",
    "C0103",
    "C0114",
    "E0401",
    "C0115",
    "C0116",
    "E1101",
    "R0801",
    "C0202",
    "W0406",
    "W1514",
    "W0201",
    "W0511",
]

[tool.pylint.main]
load-plugins = "pylint_actions, perflint"
